requirements.txt
sentence-transformers
faiss-cpu
tqdm
lxml
beautifulsoup4
numpy
scikit-learn
streamlit  # optional
